Ursula. Case 28191. Call to container move.

+++ NbtWebApp/app/composites/csw.grid.js
@@ -243,12 +243,14 @@
                     gridopts.selType = 'checkboxmodel';
                     gridopts.selModel = { mode: 'Simple' };
                     gridopts.listeners.selectionchange = function (t, selected, eOpts) {
-                        if (Csw.isNullOrEmpty(selected) || selected.length === 0) {
-                            cswPrivate.editAllButton.disable();
-                            cswPrivate.deleteAllButton.disable();
-                        } else {
-                            cswPrivate.editAllButton.enable();
-                            cswPrivate.deleteAllButton.enable();
+                        if(cswPrivate.editAllButton && cswPrivate.deleteAllButton) {
+                            if (Csw.isNullOrEmpty(selected) || selected.length === 0) {
+                                cswPrivate.editAllButton.disable();
+                                cswPrivate.deleteAllButton.disable();
+                            } else {
+                                cswPrivate.editAllButton.enable();
+                                cswPrivate.deleteAllButton.enable();
+                            }
                         }
                     };
                 } else {
